% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_S3_Generics.R, R/color_palettes.R
\name{checkColorMap}
\alias{checkColorMap}
\alias{checkColorMap.default}
\title{Check color map for a factor}
\usage{
checkColorMap(x, colors = NULL, ...)

\method{checkColorMap}{default}(x, colors = NULL, ...)
}
\arguments{
\item{x}{An R object contains the factor vector to be checked.}

\item{colors}{A named vector, whose names are factor levels and values are
colors. If is \code{NULL}, just generate colors with \code{setColor}.
Otherwise, it first generates colors for each level, then replaces those with
names mapping to \code{colors}.}

\item{...}{Arguments passed to other methods.}
}
\value{
An updated \code{colors} vector, whose names are identical to the
levels.
}
\description{
Function to make sure that all levels of a factor map to distinct colors.
}
\examples{
## Assign colors for a character or factor vector.
xx <- sample(LETTERS, 10)
cols <- setColor(xx)

## Ensure each level in 'xx' get a color
checkColorMap(xx, cols)

}
