% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkSameLength}
\alias{checkSameLength}
\title{Check whether the lengths of input objects are equal}
\usage{
checkSameLength(...)
}
\arguments{
\item{...}{R objects to be compared}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check whether the lengths of input objects are equal
}
