% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_ops.R
\name{fastIntersect}
\alias{fastIntersect}
\title{A fast version of base::intersect()}
\usage{
fastIntersect(x, y, keep.duplicated = FALSE)
}
\arguments{
\item{x, y}{Vectors to be compared.}

\item{keep.duplicated}{Whether or not to keep duplicated elements in \code{x}}
}
\value{
A vector of a common mode.
}
\description{
A fast version of base::intersect()
}
\examples{
x <- sample(LETTERS, 12)
y <- sample(LETTERS, 12)
fastIntersect(x, y)

}
\references{
\url{https://stackoverflow.com/questions/72631297/speed-up-
setdiff-intersect-union-operations-on-vectors-in-r}
}
\seealso{
\code{\link[base]{intersect}}
}
