% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{fetchColnames}
\alias{fetchColnames}
\title{Fetch column names exists in the data object}
\usage{
fetchColnames(object, query)
}
\arguments{
\item{object}{Any object that has implemented \code{colnames(object)}.}

\item{query}{Column names to check.}
}
\description{
Fetch column names exists in the data object
}
