% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{isValidCharacters}
\alias{isValidCharacters}
\title{Check valid characters}
\usage{
isValidCharacters(x)
}
\arguments{
\item{x}{A vector, matrix or list}
}
\value{
A logical vector
}
\description{
Check if input characters are valid (neither \code{NA} nor \code{""})
}
\examples{
isValidCharacters(c("a", "", "b"))
isValidCharacters(c("a", NA, "b"))

}
