% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{pal_discrete}
\alias{pal_discrete}
\title{A palette list for distinct colors.}
\format{
An object of class \code{list} of length 20.
}
\usage{
pal_discrete
}
\description{
A palette list for distinct colors.
}
\keyword{datasets}
