% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{easy.utils-package}
\alias{easy.utils}
\alias{easy.utils-package}
\title{easy.utils: Frequently Used Functions for Easy R Programming}
\description{
Some utility functions for validation and data manipulation. These functions can be helpful to reduce internal codes everywhere in package development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ycli1995/easy.utils}
  \item Report bugs at \url{https://github.com/ycli1995/easy.utils/issues}
}

}
\author{
\strong{Maintainer}: Yuchen Li \email{ycli1995@outlook.com}

}
\keyword{internal}
