% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{fetchColnames}
\alias{fetchColnames}
\title{Fetch column names exists in the data object}
\usage{
fetchColnames(object, query)
}
\arguments{
\item{object}{Any object that has implemented \code{colnames(object)}.}

\item{query}{Column names to check.}
}
\value{
An update \code{query} where only entries existing in 
\code{colnames(object)} are kept. If no any \code{query} was found, raise an 
error.
}
\description{
Fetch column names exists in the data object
}
